fix: Make search work in a more predictable way
Article search was implemented using a text index for both title and content in
the MongoDB database used by the app, but it didn't work as I expected it to
(e.g. passing part of a word as a query didn't give any results, since it is
more complex than a regular substring search).

Now it only searches by title using a simple case insensitive regex.